# Bit-exact version of Lab ⇔ RGB conversions