… modules/core/src/softfloat.cpp
… modules/imgproc/src/color.cpp