### This pullrequest changes
1. Adding new version of Lab2RGB_b conversion which is done in integers. Other Lab to/from RGB conversions are already bit-exact or performed on float input data.

2. Lookup-tables for cube root, gamma correction and coefficients for XYZ ⇔ RGB conversion are calculated using SoftFloat. N.B. As a result RGB2Lab_b produces slightly different values than before, see the accuracy paragraph below. Luv conversions and OpenCL version of Lab conversions are also affected.

3. `softfloat.cpp`: internal functions made static, extra functions removed.

After these changes Lab to/from RGB conversion should produce the same results on all machines regardless of architecture, float point implementation, etc.

N.B. RGB2Lab_f and trilinear interpolation code is moved to rgb2lab_faster branch since it has nothing to do with bit exactness.
N.B. All internal accuracy & speed tests are moved to `lab_tetra` branch.

### Performance

Measured 100 runs on random 257x257 image
Repeated 100 times and then averaged

Version | Linux x64 Core i5-6600 @ 3.3 GHz |  Win x64 Core i5-6300U @ 2.4 GHz
------------ | ------------- | -------------
Bit-exact | 0.055 s | 0.080 s
Original | 0.045 s | 0.068 s

### Accuracy
Max value per channel is 255

Lab2RGB_b:
- max error: 2
- mean by BGR channels: [0.018, 0.033, 0.019]
- stddev by BGR channels: [0.133, 0.172, 0.132]

RGB2Lab_b:
- max error: 1
- mean by Lab channels: [0.001 0.000 0.000]
- stddev by Lab channels: [0.032 0.003 0.001]

Lookup tables max error, avg error:
- LabCbrtTab: 1.19209e-07, 8.79e-09
- LabCbrtTab_b: 1, 6.51e-04
- sRGBGammaTab: 5.11e-10, 6.22e-14
- sRGBInvGammaTab: 8.18e-09, 6.6e-13
- other tables: 0, 0