@savuor 2017-06-21
@vpisarev Performance problems fixed (see the table), now we can merge it