@alalek 2017-07-14
I have squashed this PR and resolved merge conflicts.
Original commits are here: https://github.com/alalek/opencv/commits/pr8498