@savuor 2017-07-15
@alalek I've fixed the conflicts and squashed the commits again, please take a look